msm: smp2p: add SSR boot-up handshake
Booting up after an SSR, a remote subsystem will go through the
following sequence if the SMP2P is already in the OPENED state:
	1) write 0 to all entries
	2) trigger an interrupt to the remote processor (which may be Apps)
	3) clients may now start writing to entries

If Apps does not process the interrupt in step 2 until after the client
writes to an entry (step 3), then Apps will not notify clients of the
X->0 transition from step 1.

Add an SSR handshake to ensure that the remote processor has time to
process the interrupt from step 2.

CRs-Fixed: 528683
Change-Id: Ia04324c2aee38c907cded2c742b809946ec4040a
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>