msm: mdss: fix frame done not signaling for command mode
When ping pong done schedule_work is used to schedule a workqueue to
signal the frame done signaling. Schedule work does not guarantee one to
one mapping to the signals done and if a second interrupt comes before
workqueue is done it may lead to some frame done signaling not
happening. Fix this by having additional count for the frames done.

Change-Id: Ie0a67ab45657ff6ef894f6958e9e4f1309ad9455
Signed-off-by: Adrian Salido-Moreno <adrianm@codeaurora.org>